XRENDERING-592: Unsafe production of XMLReader objects
Sets the FEATURE_SECURE_PROCESSING feature in the SAXParserFactory. The classes could be modified to retrieve the XMLReader from ComponentManager instead, but I'm not sure of the possible side-effects of that.

In AbstractStreamParser, it also sets an EntityResolver so DTDs can be found (the XMLReader no longer retrieves everything).
Breakages arising from this change could occur, if a processed document is using an unsafe DOCTYPE declaration.