XRENDERING-705: Footnote macro produces inline content in standalone mode
* Wrap the footnote content in a paragraph when the macro is standalone.
* Adapt tests to expect the wrapping paragraph.

While this fixes the immediate problem of invalid HTML, it isn't nice to
have a standalone footnote macro in the WYSIWYG editor. To fix this, a
possible solution tracked in XRENDERING-784 is that macros should have a
way to indicate their preferred mode (inline or standalone), and the
WYSIWYG editor should honor this preferred mode.